# cpu/nrf52: enable instruction cache